+++ web/src/search/panels/LoadingPanelView.tsx
